New installer (#122)
* Support the new install for perf testing

* Fix servicegraph

* Get prometheus working

* Remove stdio rules

* Label namespaces

* Fix sidecar

* wip

* Use other operator

* Fix load script

* Enable kubelet, pin operator version

+++ perf/istio-install/setup_prometheus.sh
@@ -7,26 +7,29 @@ mkdir -p "${WD}/tmp"
 
 function install_prometheus() {
   local DIRNAME="$1" # should be like tools/perf/istio/tmp
-  local PROMETHEUS_INSTALL="${DIRNAME}/../../prometheus-install"
+  local PROMETHEUS_INSTALL="${DIRNAME}"
 
   # Create GCP SSD Storage class for Prometheus to use. May not work outside GKE
-  kubectl apply -f "${PROMETHEUS_INSTALL}/ssd-storage-class.yaml"
+  kubectl apply -f "${WD}/../prometheus-install/ssd-storage-class.yaml"
 
   helm fetch stable/prometheus-operator --untar --untardir "${PROMETHEUS_INSTALL}"
 
   kubectl create ns istio-prometheus || true
-  PROM_OP=tmp/prom-operator.yaml
+  PROM_OP="$DIRNAME/prom-operator.yaml"
   helm template "${PROMETHEUS_INSTALL}/prometheus-operator/" \
     --namespace istio-prometheus \
     --name prometheus \
-    -f "${PROMETHEUS_INSTALL}/values.yaml" > "${PROM_OP}"
+    -f "${WD}/../prometheus-install/values.yaml" > "${PROM_OP}"
 
   echo "${PROM_OP}"
 
-  kubectl apply --namespace istio-prometheus -f "${PROM_OP}"
 
-  # delete grafana pod so it redeploys with new config
-  kubectl delete pod -l app=grafana
+  if [[ -z "${DRY_RUN}" ]]; then
+    kubectl apply --namespace istio-prometheus -f "${PROM_OP}"
+
+    # delete grafana pod so it redeploys with new config
+    kubectl delete pod -l app=grafana
+  fi
 }
 
 install_prometheus $1
